Scaffold Hello World Backend
Introduces a new backend type service to work in conjunction with
'./backend'. Currently, the idea is to make this application do the
brunt of the work, but to allow incremental improvement by proxying
requests to './backend' for non-existing features. For a full run
down of this type of development check out the article:

https://www.parsonsmatt.org/2016/07/08/servant-persistent_updated.html