Initial Login Scaffold
This sets up the initial framework for gaining authorization from google and
persisting said token. It does not yet deal with prolonged authentication, or
anything of that nature, however it provides a good scaffold. As of yet it is
untested (aside from working as a web application) and hence should not be
considered example code until a later revision.