Working Log In Flow With User Details
This commit finalizes the login flow (at least with regards to Google, though
other providers will be similar enough to avoid duplication). We can now

1. Get an initial token from Google
2. Exchange this token for a request token
3. Extract user details from this request and store them

This allows us to get the user id and store it for future sessions, effectively
differentiating logging in and creating a new user, a previous failing.