Implement Authentication Part I
In part I of authentication, we authenticate our external API, and set the
template for all future external API's. We also setup JWT and Cookie (session)
based authentication for our entire API. In part II we will complete the hookup
and test vigorously with the application. However, with part II it should be
noted that the core of the application will then be complete, with only light
refactoring needed to add all remaining servants and hit MINIMUM VIABLE PRODUCT!